Consolidate spawnProc logic
Use separate build files only for the parts of these functions that
differ, which allows us to consolidate the logic and reduces the
chance of errors.
