+++ proc.go
@@ -2,10 +2,52 @@ package main
 
 import (
 	"errors"
+	"fmt"
 	"os"
+	"os/exec"
 	"sync"
 	"time"
 )
+
+// spawnProc starts the specified proc, and returns any error from running it.
+func spawnProc(proc string, errCh chan<- error) {
+	procObj := procs[proc]
+	logger := createLogger(proc, procObj.colorIndex)
+
+	cs := append(cmdStart, procObj.cmdline)
+	cmd := exec.Command(cs[0], cs[1:]...)
+	cmd.Stdin = nil
+	cmd.Stdout = logger
+	cmd.Stderr = logger
+	cmd.SysProcAttr = procAttrs
+
+	cmd.Env = append(os.Environ(), fmt.Sprintf("PORT=%d", procObj.port))
+
+	fmt.Fprintf(logger, "Starting %s on port %d\n", proc, procObj.port)
+	if err := cmd.Start(); err != nil {
+		select {
+		case errCh <- err:
+		default:
+		}
+		fmt.Fprintf(logger, "Failed to start %s: %s\n", proc, err)
+		return
+	}
+	procObj.cmd = cmd
+	procObj.stoppedBySupervisor = false
+	procObj.mu.Unlock()
+	err := cmd.Wait()
+	procObj.mu.Lock()
+	procObj.cond.Broadcast()
+	if err != nil && procObj.stoppedBySupervisor == false {
+		select {
+		case errCh <- err:
+		default:
+		}
+	}
+	procObj.waitErr = err
+	procObj.cmd = nil
+	fmt.Fprintf(logger, "Terminating %s\n", proc)
+}
 
 // Stop the specified proc, issuing os.Kill if it does not terminate within 10
 // seconds. If signal is nil, os.Interrupt is used.
